# changed x and y position to floating points type

+++ headers/UI/UI_Element.h
@@ -8,9 +8,9 @@
 typedef struct UI_Element
 {
     //x position
-    int x;
+    float x;
     //y position
-    int y;
+    float y;
     //width
     int width;
     //height
@@ -26,7 +26,7 @@ typedef struct UI_Element
 } UI_Element;
 
 //Set the fields of element to the values passed in
-UI_Element *UI_Element_set_fields(UI_Element *element, int x, int y, int width, int height, Outline outline, Inline inline_, float scale, int anchor);
+UI_Element *UI_Element_set_fields(UI_Element *element, float x, float y, int width, int height, Outline outline, Inline inline_, float scale, int anchor);
 
 //Set the size of element to the size of frame if it's smaller
 //!! Do not put a label rect for element, only for frame if necessary !!
@@ -34,9 +34,9 @@ UI_Element *UI_Element_set_fields(UI_Element *element, int x, int y, int width, 
 void UI_Element_match_size(UI_Element *element, UI_Element *frame, float scale_x, float scale_y);
 
 //return true if (x, y) is in element, false otherwise
-int UI_element_collision(UI_Element *element, int x, int y, float scale_x, float scale_y);
+int UI_element_collision(UI_Element *element, float x, float y, float scale_x, float scale_y);
 //Positions the element at (x, y) based on the anchor and the scales
-void UI_Element_set_position(UI_Element *element, int x, int y, float x_position_factor, float y_position_factor, float scale_x, float scale_y, int anchor);
+void UI_Element_set_position(UI_Element *element, float x, float y, float x_position_factor, float y_position_factor, float scale_x, float scale_y, int anchor);
 
 //Draw the element inline
 void UI_Element_render_inline(UI_Element *element, float scale_x, float scale_y);
